Extracted client card details into a component
