+++ src/components/Clients/ClientCard/index.tsx
@@ -1,5 +1,4 @@
 import React, { useState } from "react";
-import { Link } from "react-router-dom";
 import { useTheme } from "@mui/material/styles";
 import Stack from "@mui/material/Stack";
 import Collapse from "@mui/material/Collapse";
@@ -8,11 +7,9 @@ import ExpandMoreOutlined from "@mui/icons-material/ExpandMoreOutlined";
 import ExpandLessOutlined from "@mui/icons-material/ExpandLessOutlined";
 
 import { ClientData } from "../../../api";
-import { useAppDispatch } from "../../../redux/hooks";
-import { setBookingsFormFrom, setBookingsFormName, setBookingsFormTo } from "../../../redux/bookingsFormSlice";
 
 import M3IconButton from "../../m3/M3IconButton";
-import M3TextButton from "../../m3/M3TextButton";
+import Details from "./Details";
 
 type ClientCardProps = {
   client: ClientData
@@ -20,12 +17,7 @@ type ClientCardProps = {
 
 export default function ClientCard({ client }: ClientCardProps): JSX.Element {
   const theme = useTheme();
-  const dispatch = useAppDispatch();
   const [open, setOpen] = useState(false);
-
-  const formattedFrom = (new Date(client.booking.from)).toLocaleDateString();
-  const formattedTo = (new Date(client.booking.to)).toLocaleDateString();
-  const periodStr = `${formattedFrom} - ${formattedTo}`;
 
   return (
     <Stack sx={{
@@ -48,27 +40,7 @@ export default function ClientCard({ client }: ClientCardProps): JSX.Element {
         </M3IconButton>
       </Stack>
       <Collapse in={open} easing={theme.transitions.easing.fastOutSlowIn}>
-        <Stack spacing={2} sx={{
-          p: "1rem",
-          borderTop: `1px solid ${theme.palette.outline.light}`
-        }}>
-          <Typography variant="titleLarge">Prenotazione</Typography>
-          <Stack sx={{ pl: "1rem" }}>
-            <Typography variant="titleMedium">{client.booking.name}</Typography>
-            <Typography variant="bodySmall">{periodStr}</Typography>
-          </Stack>
-          <Stack alignItems="flex-end">
-            <Link to={`/bookings/${client.booking.id}`} style={{ textDecoration: "none" }}>
-              <M3TextButton onClick={() => {
-                dispatch(setBookingsFormFrom(client.booking.from));
-                dispatch(setBookingsFormTo(client.booking.to));
-                dispatch(setBookingsFormName(client.booking.name));
-              }}>
-                Mostra prenotazione
-              </M3TextButton>
-            </Link>
-          </Stack>
-        </Stack>
+        <Details client={client} />
       </Collapse>
     </Stack>
   );
